ADDED:UserInfo for facebook userid/access_token front2backend communication and implements on FbUser profile save to MongoDB functions

+++ Spring-boot/src/main/java/crispy_octo_moo/domain/BaseEntity.java
@@ -42,7 +42,7 @@ public abstract class BaseEntity implements Serializable {
     }
     
     @Version
-    private int version;
+    private Long version;
 
     /**
      *  All objects will have a unique UUID which allows for the decoupling from DB generated ids
@@ -92,7 +92,7 @@ public abstract class BaseEntity implements Serializable {
         return getUuid().toString();
     }
 
-    public int getVersion() {
+    public Long getVersion() {
         return version;
     }
